+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.0.0 (2025-10-23)
 
 
+### Bug Fixes
+
+* include download urls in userscript metadata (#2) 92a467f, closes #2
+
+